+++ plugins/wasm-test/src/lib.rs
@@ -5,15 +5,13 @@ use std::collections::HashMap;
 
 #[host_fn]
 extern "ExtismHost" {
-    fn trace(input: Json<TraceInput>);
+    fn host_log(input: Json<HostLogInput>);
     fn exec_command(input: Json<ExecCommandInput>) -> Json<ExecCommandOutput>;
 }
 
 #[plugin_fn]
 pub fn register_tool(_: ()) -> FnResult<Json<ToolMetadataOutput>> {
-    unsafe {
-        trace(Json("Registering tool".into()))?;
-    }
+    host_log!("Registering tool".into());
 
     Ok(Json(ToolMetadataOutput {
         name: "WASM Test".into(),